mark as watched hook seperated
